true	QuestionSetController.java	-/Modified	Commit	src\main\java\cs544\letmegiveexam\controller\QuestionSetController.java true	UserExamController.java	-/Modified	Commit	src\main\java\cs544\letmegiveexam\controller\UserExamController.java true	UserExamDAO.java	-/Modified	Commit	src\main\java\cs544\letmegiveexam\dao\UserExamDAO.java true	Question.java	-/Modified	Commit	src\main\java\cs544\letmegiveexam\model\Question.java true	QuestionSet.java	-/Modified	Commit	src\main\java\cs544\letmegiveexam\model\QuestionSet.java true	UserExam.java	-/Modified	Commit	src\main\java\cs544\letmegiveexam\model\UserExam.java true	UserExamService.java	-/Modified	Commit	src\main\java\cs544\letmegiveexam\service\UserExamService.java true	sideBar.jsp	-/Modified	Commit	src\main\webapp\WEB-INF\template\sideBar.jsp true	examHistory.jsp	-/Modified	Commit	src\main\webapp\WEB-INF\views\examHistory.jsp true	listQuestion.jsp	-/Modified	Commit	src\main\webapp\WEB-INF\views\listQuestion.jsp

+++ src/main/java/cs544/letmegiveexam/service/UserExamService.java
@@ -39,7 +39,7 @@ public class UserExamService {
     }
     
     @Transactional(propagation = Propagation.REQUIRED)
-    public List<UserExam> getUserExam(long userId) {
+    public List<UserExam> getUserExam(int userId) {
         return userExamDAO.getUserExam(userId);
     }
 }
